refactor: move http and typeorm on module customer to infra layer
