+++ src/modules/customers/services/DeleteCustomerService.ts
@@ -1,5 +1,5 @@
 import { getCustomRepository } from 'typeorm';
-import { CustomerRepository } from '../typeorm/repositories/CustomerRepository';
+import { CustomerRepository } from '../infra/typeorm/repositories/CustomerRepository';
 import { ShowCustomerService } from './ShowCustomerService';
 
 interface ICustomerRequest {
