+++ src/modules/customers/services/ShowCustomerService.ts
@@ -1,7 +1,7 @@
 import { getCustomRepository } from 'typeorm';
 import { AppError } from '@shared/errors/AppError';
-import { Customer } from '../typeorm/entities/Customer';
-import { CustomerRepository } from '../typeorm/repositories/CustomerRepository';
+import { Customer } from '../infra/typeorm/entities/Customer';
+import { CustomerRepository } from '../infra/typeorm/repositories/CustomerRepository';
 
 interface ICustomerRequest {
   customerId: string;
